Update SPARQL QUERYS for DBLP, GS, MA and SCOPUS Add SyntacticDistance class Add SemanticDistance class Add semanticdistance service: prov: Jose Ortiz ucuenca Update semanticdistanceservice cache using posgresql Add CommonsServices ( remoceAccents, cleantext, etc) renames clases
