+++ tools/commons/pom.xml
@@ -63,6 +63,7 @@
     </build>
 
     <dependencies>
+        <!-- Test dependencies -->
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
@@ -70,6 +71,14 @@
             <scope>test</scope>
             <type>jar</type>
         </dependency>
+        
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>${postgresql.version}</version>
+            <scope>test</scope>
+            <type>jar</type>
+        </dependency> 
         <!-- Test dependencies -->
  
         <dependency>
@@ -104,7 +113,18 @@
             <version>2.3.1</version>
             <type>jar</type>
         </dependency>
-
+        <!-- http://mvnrepository.com/artifact/postgresql/postgresql -->
+        <!--<dependency>
+            <groupId>postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>9.1-901-1.jdbc4</version>
+        </dependency>-->
+        <!--
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>${postgresql.version}</version>
+        </dependency>    -->
     </dependencies>
    
 </project>
