+++ tools/commons/src/main/java/org/apache/marmotta/ucuenca/wk/commons/service/ConstantService.java
@@ -65,5 +65,7 @@ public interface ConstantService {
     String getGSGraph();
     
     String getClusterGraph();
+    
+    String getAuthorsGraph();
 
 }
