+++ tools/commons/src/main/java/org/apache/marmotta/ucuenca/wk/commons/service/TranslationService.java
@@ -13,5 +13,5 @@ import com.google.gson.JsonObject;
 public interface TranslationService {
 
     JsonObject translate(String text);
-
+  
 }
