+++ tools/commons/src/main/resources/META-INF/beans.xml
